Split tests on more files
